feat:#222 칼로리 바 차트 컴포넌트 분리
